Package lock bump and test fixes
Running npm install updated the package-lock.json

Replaced the `.replaceAll` calls in `puppeteer.test.js` to `.replace` calls.
I don't know why, but Node kinda sucks sometimes and I can't find a stable node version that implements `.replaceAll`.
This seems to be a common issue: https://stackoverflow.com/questions/65295584/jest-typeerror-replaceall-is-not-a-function

So I took the shortest path and replaced to `.replace` with regex in order to avoid future compatibility issues with new contributors ~like me~